+++ src/entity/item/Item.java
@@ -4,15 +4,15 @@ import entity.Entity;
 import graphic.Sprite;
 
 public abstract class Item extends Entity {
-    public Item(int x, int y) {
+    public Item(double x, double y) {
         super(x, y);
     }
 
-    public Item(int x, int y, Sprite[] sprite) {
+    public Item(double x, double y, Sprite[] sprite) {
         super(x, y, sprite);
     }
 
-    public Item(int x, int y, Sprite[] sprite, int cur) {
+    public Item(double x, double y, Sprite[] sprite, int cur) {
         super(x, y, sprite);
         this.cur = cur;
     }
