+++ src/entity/tile/Wall.java
@@ -4,11 +4,11 @@ import graphic.Sprite;
 
 public class Wall extends Tile{
 
-    public Wall(int x, int y, Sprite[] sprite) {
+    public Wall(double x, double y, Sprite[] sprite) {
         super(x, y, sprite);
     }
 
-    public Wall(int x, int y) {
+    public Wall(double x, double y) {
         super(x, y);
     }
 }
